+++ src/Core/Factorization/Zassenhaus.hpp
@@ -23,7 +23,7 @@ ast::AST* zassenhaus(ast::AST* f, ast::AST* x);
  * @param q A prime integer p
  * @return A list of tuples, the first element of the tuple is a factor, and the second is its degree
  */
-ast::AST* cantorZassenhausDDF(ast::AST* a, ast::AST* x, long p);
+ast::AST* cantorZassenhausDDF(ast::AST* a, ast::AST* x, Int p);
 
 /**
  * @brief Given a square-free polynomial a(x) and a integer n in Zp[x], 
@@ -35,7 +35,7 @@ ast::AST* cantorZassenhausDDF(ast::AST* a, ast::AST* x, long p);
  * @param p A prime integer
  * @return A list of equal degree factors
  */
-ast::AST* cantorZassenhausEDF(ast::AST* a, ast::AST* x, long n, long p);
+ast::AST* cantorZassenhausEDF(ast::AST* a, ast::AST* x, Int n, Int p);
 
 /**
  * @brief Given an univariate polynomial in Zp[x], Compute its unique factorization in Zp[x]
@@ -45,7 +45,7 @@ ast::AST* cantorZassenhausEDF(ast::AST* a, ast::AST* x, long n, long p);
  * @param p A prime integer
  * @return A product expression with all the irreductible factors of u(x) in Zp[x] as operands
  */
-ast::AST* cantorZassenhaus(ast::AST* u, ast::AST* x, long p);
+ast::AST* cantorZassenhaus(ast::AST* u, ast::AST* x, Int p);
 
 }
